Overhauled exceptions
