+++ src/main/kotlin/de/thm/mni/ii/classroom/controller/BBBApiController.kt
@@ -4,7 +4,7 @@ import de.thm.mni.ii.classroom.model.api.CreateRoomBBB
 import de.thm.mni.ii.classroom.model.api.JoinRoomBBBResponse
 import de.thm.mni.ii.classroom.model.api.MessageBBB
 import de.thm.mni.ii.classroom.model.api.ReturnCodeBBB
-import de.thm.mni.ii.classroom.exception.ApiException
+import de.thm.mni.ii.classroom.exception.api.ApiException
 import de.thm.mni.ii.classroom.services.DownstreamApiService
 import org.slf4j.LoggerFactory
 import org.springframework.http.ResponseEntity
