--- src/main/kotlin/de/thm/mni/ii/classroom/exception/ApiException.kt
@@ -1,7 +1,0 @@
-package de.thm.mni.ii.classroom.exception
-
-class ApiException(
-    val bbbMessageKey: String = "InternalServerError",
-    cause: Throwable? = null,
-    val bbbMessage: String = "An error occurred: ${cause?.message}",
-    ): Exception(bbbMessage, cause)
