--- src/main/kotlin/de/thm/mni/ii/classroom/exception/ClassroomNotFoundException.kt
@@ -1,7 +1,0 @@
-package de.thm.mni.ii.classroom.exception
-
-class ClassroomNotFoundException(
-        meetingID: String = "",
-        private val bbbMessageKey: String = "meetingDoesNotExist",
-        private val bbbMessage: String = "MeetingID $meetingID does not exist!"
-    ): Exception(bbbMessage)
