+++ src/main/kotlin/de/thm/mni/ii/classroom/exception/api/ClassroomNotFoundException.kt
@@ -1,0 +1,7 @@
+package de.thm.mni.ii.classroom.exception.api
+
+class ClassroomNotFoundException(
+        meetingID: String = "",
+        override val bbbMessageKey: String = "meetingDoesNotExist",
+        override val bbbMessage: String = "MeetingID $meetingID does not exist!"
+    ): ApiException(bbbMessage)
