+++ src/main/kotlin/de/thm/mni/ii/classroom/exception/api/MissingMeetingIDException.kt
@@ -1,0 +1,6 @@
+package de.thm.mni.ii.classroom.exception.api
+
+class MissingMeetingIDException(
+    override val bbbMessageKey: String = "missingParamMeetingID",
+    override val bbbMessage: String = "You must specify a meeting ID for the meeting."
+): ApiException(bbbMessage)
