+++ src/main/kotlin/de/thm/mni/ii/classroom/exception/classroom/ClassroomException.kt
@@ -1,0 +1,3 @@
+package de.thm.mni.ii.classroom.exception.classroom
+
+open class ClassroomException(message: String): Exception(message)
