+++ src/main/kotlin/de/thm/mni/ii/classroom/exception/classroom/TicketAlreadyExistsException.kt
@@ -1,0 +1,5 @@
+package de.thm.mni.ii.classroom.exception.classroom
+
+import de.thm.mni.ii.classroom.model.classroom.Ticket
+
+class TicketAlreadyExistsException(ticket: Ticket): ClassroomException("Similar ticket \"${ticket.description}\" already exists!")
